+++ lib/src/data_frame/variables_extractor/variables_extractor.dart
@@ -4,5 +4,5 @@ import 'package:xrange/zrange.dart';
 abstract class VariablesExtractor {
   Matrix get features;
   Matrix get labels;
-  Set<ZRange> get categoricalIndices;
+  Set<ZRange> get encodedColumnRanges;
 }