+++ test/data_frame/variables_extractor/variables_extractor_impl_test.dart
@@ -280,7 +280,7 @@ void main() {
       final valueConverter = mocks.ToFloatNumberConverterMock();
       final extractor = VariablesExtractorImpl(data, columnIndices, rowIndices,
           encoders, labelIdx, valueConverter);
-      final actual = extractor.categoricalIndices;
+      final actual = extractor.encodedColumnRanges;
 
       expect(actual, equals([
         ZRange.closed(0, 2),